elyazısı_dogruluk/requirements.txt
flask==2.2.5
Werkzeug<3.1
tensorflow==2.15.0
pillow==10.1.0
numpy==1.24.3
FuzzyTM>=0.4.0
psutil>=5.9.1